requirements.txt
apex==0.1
matplotlib==3.4.2
opencv-python==4.5.3.56
pandas==1.3.1
protobuf==3.17.3
PyYAML==5.4.1
ruamel.yaml==0.17.10
scipy==1.7.0
tensorboard==2.5.0
termcolor==1.1.0
thop==0.0.31.post2005241907
torch==1.8.1
torchstat==0.0.7
torchvision==0.9.1
tqdm==4.61.2
numpy==1.20.3